App/VQC/requirements.txt
waitress
Flask==1.1.2
pymongo==3.12.0
requests==2.25.1
Flask_PyMongo==2.3.0
opencv_python==4.5.3.56
Werkzeug==1.0.1
numpy==1.20.3
pytest==6.2.3
confuse==1.5.0
PyYAML==5.4.1
itsdangerous==2.0.1
